+++ CHANGELOG.md
@@ -27,6 +27,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - pre_transform filter to KiCad variants
 - E/DRC:
   - Dark mode support and icon to select it.
+- Blender Export:
+  - Added support for PCB2Blender 2.17 format, and made default
 - BoM:
   - JSON output format. As a helper for external renderers (See #917)
   - Dark mode support and icon to select it.